refactor(gateway): isolate auth logout helpers
