+++ src/gateway/server-methods/models-auth-logout.ts
@@ -1,0 +1,70 @@
+import {
+  removeProviderAuthProfilesWithLock,
+  resolvePersistedAuthProfileOwnerAgentDir,
+} from "../../agents/auth-profiles.js";
+import type { ChatAbortOps } from "../chat-abort.js";
+import type { GatewayRequestContext } from "./types.js";
+
+type LogoutProfileSelection = { ok: true; profileIds?: string[] } | { ok: false; message: string };
+
+export function readLogoutProfileSelection(
+  params: Record<string, unknown>,
+): LogoutProfileSelection {
+  if (!("profileIds" in params)) {
+    return { ok: true };
+  }
+  if (!Array.isArray(params.profileIds) || params.profileIds.length === 0) {
+    return { ok: false, message: "profileIds must be a non-empty string array" };
+  }
+  const profileIds: string[] = [];
+  for (const value of params.profileIds) {
+    if (typeof value !== "string" || !value.trim()) {
+      return { ok: false, message: "profileIds must be a non-empty string array" };
+    }
+    const profileId = value.trim();
+    if (!profileIds.includes(profileId)) {
+      profileIds.push(profileId);
+    }
+  }
+  return { ok: true, profileIds };
+}
+
+export function createAuthLogoutAbortOps(context: GatewayRequestContext): ChatAbortOps {
+  return {
+    chatAbortControllers: context.chatAbortControllers,
+    chatRunState: context.chatRunState,
+    removeChatRun: context.removeChatRun,
+    agentRunSeq: context.agentRunSeq,
+    broadcast: context.broadcast,
+    nodeSendToSession: context.nodeSendToSession,
+  };
+}
+
+// Auth profiles can be adopted by a provider-specific owner agent dir. Logout
+// must remove every owning store or stale profiles reappear on the next status
+// read and provider-auth warmup.
+export async function removeProviderAuthProfilesAcrossOwnerStores(params: {
+  provider: string;
+  agentDir: string;
+  profileIds: string[];
+}): Promise<boolean> {
+  const ownerAgentDirs = new Set<string | undefined>([params.agentDir]);
+  for (const profileId of params.profileIds) {
+    ownerAgentDirs.add(
+      resolvePersistedAuthProfileOwnerAgentDir({
+        agentDir: params.agentDir,
+        profileId,
+      }),
+    );
+  }
+  for (const ownerAgentDir of ownerAgentDirs) {
+    const updatedStore = await removeProviderAuthProfilesWithLock({
+      provider: params.provider,
+      agentDir: ownerAgentDir,
+    });
+    if (!updatedStore) {
+      return false;
+    }
+  }
+  return true;
+}
