+++ src/gateway/server-methods/models-auth-status.ts
@@ -23,7 +23,6 @@ import {
   externalCliDiscoveryForConfigStatus,
   listProfilesForProvider,
   removeAuthProfilesAcrossOwnerStores,
-  removeProviderAuthProfilesWithLock,
   resolvePersistedAuthProfileOwnerAgentDir,
   setAuthProfileOrder,
 } from "../../agents/auth-profiles.js";
@@ -52,9 +51,14 @@ import { providerUsageLabel, resolveUsageProviderId } from "../../infra/provider
 import type { UsageProviderId } from "../../infra/provider-usage.types.js";
 import { createSubsystemLogger } from "../../logging/subsystem.js";
 import { refreshActiveProviderAuthRuntimeSnapshot } from "../../secrets/runtime.js";
-import { abortChatRunsForProvider, type ChatAbortOps } from "../chat-abort.js";
+import { abortChatRunsForProvider } from "../chat-abort.js";
 import { formatForLog } from "../ws-log.js";
 import { modelAuthAgentScopeError, resolveModelAuthAgentScope } from "./model-auth-agent-scope.js";
+import {
+  createAuthLogoutAbortOps,
+  readLogoutProfileSelection,
+  removeProviderAuthProfilesAcrossOwnerStores,
+} from "./models-auth-logout.js";
 import {
   clearModelAuthStatusUsageCache,
   fingerprintProviderUsageCredentials,
@@ -67,7 +71,7 @@ import type {
   ModelAuthStatusProvider,
   ModelAuthStatusResult,
 } from "./models-auth-status.types.js";
-import type { GatewayRequestContext, GatewayRequestHandlers } from "./types.js";
+import type { GatewayRequestHandlers } from "./types.js";
 
 export type {
   ModelAuthExpiry,
@@ -120,68 +124,6 @@ function readProviderParam(params: Record<string, unknown>): string | null {
   }
   const provider = normalizeProviderId(raw);
   return provider || null;
-}
-
-type LogoutProfileSelection = { ok: true; profileIds?: string[] } | { ok: false; message: string };
-
-function readLogoutProfileSelection(params: Record<string, unknown>): LogoutProfileSelection {
-  if (!("profileIds" in params)) {
-    return { ok: true };
-  }
-  if (!Array.isArray(params.profileIds) || params.profileIds.length === 0) {
-    return { ok: false, message: "profileIds must be a non-empty string array" };
-  }
-  const profileIds: string[] = [];
-  for (const value of params.profileIds) {
-    if (typeof value !== "string" || !value.trim()) {
-      return { ok: false, message: "profileIds must be a non-empty string array" };
-    }
-    const profileId = value.trim();
-    if (!profileIds.includes(profileId)) {
-      profileIds.push(profileId);
-    }
-  }
-  return { ok: true, profileIds };
-}
-
-function createAuthLogoutAbortOps(context: GatewayRequestContext): ChatAbortOps {
-  return {
-    chatAbortControllers: context.chatAbortControllers,
-    chatRunState: context.chatRunState,
-    removeChatRun: context.removeChatRun,
-    agentRunSeq: context.agentRunSeq,
-    broadcast: context.broadcast,
-    nodeSendToSession: context.nodeSendToSession,
-  };
-}
-
-// Auth profiles can be adopted by a provider-specific owner agent dir. Logout
-// must remove every owning store or stale profiles reappear on the next status
-// read and provider-auth warmup.
-async function removeProviderAuthProfilesAcrossOwnerStores(params: {
-  provider: string;
-  agentDir: string;
-  profileIds: string[];
-}): Promise<boolean> {
-  const ownerAgentDirs = new Set<string | undefined>([params.agentDir]);
-  for (const profileId of params.profileIds) {
-    ownerAgentDirs.add(
-      resolvePersistedAuthProfileOwnerAgentDir({
-        agentDir: params.agentDir,
-        profileId,
-      }),
-    );
-  }
-  for (const ownerAgentDir of ownerAgentDirs) {
-    const updatedStore = await removeProviderAuthProfilesWithLock({
-      provider: params.provider,
-      agentDir: ownerAgentDir,
-    });
-    if (!updatedStore) {
-      return false;
-    }
-  }
-  return true;
 }
 
 // UI expiry fields are emitted only when both timestamp and remaining duration
